# Reduce arrays when we have throw-away return values