These arrays can really add up. When I designed this API, I didn't know that `return a,b` makes an Array 😖