@rmosolgo 2019-01-30
Oh, regarding benchmarks, it might be a hair faster: 

## Before 

```
Warming up --------------------------------------
validate - introspection 
                        39.000  i/100ms
validate - abstract fragments
                        79.000  i/100ms
validate - abstract fragments 2
                        42.000  i/100ms
validate - big query     4.000  i/100ms
Calculating -------------------------------------
validate - introspection 
                        400.885  (± 3.7%) i/s -      2.028k in   5.066375s
validate - abstract fragments
                        803.650  (± 3.1%) i/s -      4.029k in   5.018186s
validate - abstract fragments 2
                        425.843  (± 3.8%) i/s -      2.142k in   5.036916s
validate - big query     44.884  (± 4.5%) i/s -    224.000  in   5.003798s
```

## After 

```
Warming up --------------------------------------
validate - introspection 
                        41.000  i/100ms
validate - abstract fragments
                        84.000  i/100ms
validate - abstract fragments 2
                        43.000  i/100ms
validate - big query     4.000  i/100ms
Calculating -------------------------------------
validate - introspection 
                        411.762  (± 4.9%) i/s -      2.091k in   5.090332s
validate - abstract fragments
                        833.529  (± 3.6%) i/s -      4.200k in   5.045434s
validate - abstract fragments 2
                        437.908  (± 4.3%) i/s -      2.193k in   5.018191s
validate - big query     45.068  (± 4.4%) i/s -    228.000  in   5.068191s
```

(It's within the margin of error)